+++ packages/interweave/src/Interweave.js
@@ -12,10 +12,11 @@ import Filter from './Filter';
 import Matcher from './Matcher';
 import Parser from './Parser';
 import Element from './components/Element';
+import { EmojiContextShape } from './shapes';
 
-type BeforeParseCallback = (content: string, props: Object) => string;
+type BeforeParseCallback = (content: string, props: Object, context: Object) => string;
 
-type AfterParseCallback = (content: React$Node[], props: Object) => React$Node[];
+type AfterParseCallback = (content: React$Node[], props: Object, context: Object) => React$Node[];
 
 type InterweaveProps = {
   content: string,
@@ -34,6 +35,10 @@ type InterweaveProps = {
 };
 
 export default class Interweave extends React.Component<InterweaveProps> {
+  static contextTypes = {
+    emoji: EmojiContextShape,
+  };
+
   static propTypes = {
     content: PropTypes.string,
     disableFilters: PropTypes.bool,
@@ -96,7 +101,7 @@ export default class Interweave extends React.Component<InterweaveProps> {
 
     // Trigger before callbacks
     const markup = beforeCallbacks.reduce((string, callback) => {
-      const nextString = callback(string, this.props);
+      const nextString = callback(string, this.props, this.context);
 
       if (__DEV__) {
         if (typeof nextString !== 'string') {
@@ -112,7 +117,7 @@ export default class Interweave extends React.Component<InterweaveProps> {
 
     // Trigger after callbacks
     const nodes = afterCallbacks.reduce((parserNodes, callback) => {
-      const nextNodes = callback(parserNodes, this.props);
+      const nextNodes = callback(parserNodes, this.props, this.context);
 
       if (__DEV__) {
         if (!Array.isArray(nextNodes)) {