+++ packages/interweave/tests/components/Emoji.test.js
@@ -7,32 +7,39 @@ import { VALID_EMOJIS } from '../mocks';
 describe('components/Emoji', () => {
   const [[hexcode, unicode, shortcode, emoticon]] = VALID_EMOJIS;
   const shortcodes = ':enraged:, :pout:';
+  const context = {
+    emoji: {
+      compact: false,
+      locale: 'en',
+      version: 'latest',
+    },
+  };
 
   it('errors if no emoticon, shortcode or unicode', () => {
-    expect(() => shallow(<Emoji />))
+    expect(() => shallow(<Emoji />, { context }))
       .toThrowError('Emoji component requires a `unicode` character, `emoticon`, or a `shortcode`.');
   });
 
   it('returns value for invalid emoticon', () => {
-    const wrapper = shallow(<Emoji emoticon="0P" />);
+    const wrapper = shallow(<Emoji emoticon="0P" />, { context });
 
     expect(wrapper.prop('children')).toBe('0P');
   });
 
   it('returns value for invalid shortcode', () => {
-    const wrapper = shallow(<Emoji shortcode="fake" />);
+    const wrapper = shallow(<Emoji shortcode="fake" />, { context });
 
     expect(wrapper.prop('children')).toBe('fake');
   });
 
   it('returns empty for invalid unicode', () => {
-    const wrapper = shallow(<Emoji unicode="fake" />);
+    const wrapper = shallow(<Emoji unicode="fake" />, { context });
 
     expect(wrapper.prop('children')).toBe('fake');
   });
 
   it('renders with only the emoticon', () => {
-    const wrapper = shallow(<Emoji emoticon={emoticon} />);
+    const wrapper = shallow(<Emoji emoticon={emoticon} />, { context });
 
     expect(wrapper.prop('data-emoticon')).toBe(emoticon);
     expect(wrapper.prop('data-unicode')).toBe(unicode);
@@ -42,7 +49,7 @@ describe('components/Emoji', () => {
   });
 
   it('renders with only the shortcode', () => {
-    const wrapper = shallow(<Emoji shortcode={shortcode} />);
+    const wrapper = shallow(<Emoji shortcode={shortcode} />, { context });
 
     expect(wrapper.prop('data-emoticon')).toBe(emoticon);
     expect(wrapper.prop('data-unicode')).toBe(unicode);
@@ -52,7 +59,7 @@ describe('components/Emoji', () => {
   });
 
   it('renders with only the unicode', () => {
-    const wrapper = shallow(<Emoji unicode={unicode} />);
+    const wrapper = shallow(<Emoji unicode={unicode} />, { context });
 
     expect(wrapper.prop('data-emoticon')).toBe(emoticon);
     expect(wrapper.prop('data-unicode')).toBe(unicode);
@@ -62,7 +69,7 @@ describe('components/Emoji', () => {
   });
 
   it('renders with both', () => {
-    const wrapper = shallow(<Emoji shortcode={shortcode} unicode={unicode} />);
+    const wrapper = shallow(<Emoji shortcode={shortcode} unicode={unicode} />, { context });
 
     expect(wrapper.prop('data-unicode')).toBe(unicode);
     expect(wrapper.prop('data-hexcode')).toBe(hexcode);
@@ -77,7 +84,7 @@ describe('components/Emoji', () => {
         unicode={unicode}
         emojiPath="http://foo.com/path/to/{{hexcode}}.svg"
       />
-    ));
+    ), { context });
 
     expect(wrapper.find('img').prop('alt')).toBe(unicode);
     expect(wrapper.find('img').prop('src')).toBe(`http://foo.com/path/to/${hexcode}.svg`);
@@ -90,7 +97,7 @@ describe('components/Emoji', () => {
         unicode={unicode}
         emojiPath={hex => `http://foo.com/path/to/${hex.toLowerCase()}.svg`}
       />
-    ));
+    ), { context });
 
     expect(wrapper.find('img').prop('alt')).toBe(unicode);
     expect(wrapper.find('img').prop('src')).toBe(`http://foo.com/path/to/${hexcode.toLowerCase()}.svg`);
@@ -107,13 +114,13 @@ describe('components/Emoji', () => {
         emojiSize={2}
         enlargeEmoji
       />
-    ));
+    ), { context });
 
     expect(wrapper.find('img').prop('src')).toBe(`http://foo.com/path/to/4/${hexcode.toLowerCase()}.svg`);
   });
 
   it('adds class names', () => {
-    const wrapper = shallow(<Emoji shortcode={shortcode} unicode={unicode} />);
+    const wrapper = shallow(<Emoji shortcode={shortcode} unicode={unicode} />, { context });
 
     expect(wrapper.prop('className')).toBe('interweave__emoji');
 
@@ -125,7 +132,7 @@ describe('components/Emoji', () => {
   });
 
   it('sets styles when size is defined', () => {
-    const wrapper = shallow(<Emoji shortcode={shortcode} unicode={unicode} />);
+    const wrapper = shallow(<Emoji shortcode={shortcode} unicode={unicode} />, { context });
 
     expect(wrapper.prop('style')).toEqual({});
 
@@ -152,15 +159,15 @@ describe('components/Emoji', () => {
   });
 
   it('can customize large size', () => {
-    const wrapper = shallow(
+    const wrapper = shallow((
       <Emoji
         shortcode={shortcode}
         unicode={unicode}
         emojiSize={2}
         emojiLargeSize={5}
         enlargeEmoji
-      />,
-    );
+      />
+    ), { context });
 
     expect(wrapper.prop('style')).toEqual({
       display: 'inline-block',